refactor:task-005:フォルダ名を追加・編集に対応
